Fix: Reworked and removed offline_API_wrapper.js (fixes #260) (#261)
