+++ required/index.html
@@ -5,10 +5,9 @@
       window.ADAPT_BUILD_TYPE = '@@build.type';
       window.ISCOOKIELMS = false;
     </script>
-    <script src="offline_API_wrapper.js"></script>
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-    <title>Adapt</title>
+    <title></title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link href="adapt.css" type="text/css" rel="stylesheet">
     <script src="libraries/modernizr.js"></script>
@@ -18,7 +17,7 @@
   <body>
     <div id="app">
       <div id="wrapper">
-        <!-- Menu / page render here -->
+        <!-- Adapt menu / page render here -->
       </div>
     </div>
   </body>
