+++ required/index_lms.html
@@ -3,10 +3,11 @@
   <head>
     <script>
       window.ADAPT_BUILD_TYPE = '@@build.type';
+      window.ISCOOKIELMS = null;
     </script>
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-    <title>Adapt</title>
+    <title></title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link href="adapt.css" type="text/css" rel="stylesheet">
     <script src="libraries/modernizr.js"></script>
@@ -14,11 +15,11 @@
   </head>
 
   <body>
-  <div id="app">
-     <div id="wrapper">
-      <!-- Adapt menu / page renders here -->
-     </div>
-  </div>
+    <div id="app">
+      <div id="wrapper">
+        <!-- Adapt menu / page renders here -->
+      </div>
+    </div>
   </body>
 
 </html>
