+++ required/scorm_test_harness.html
@@ -1,16 +1,15 @@
 <!doctype html>
-<html id="adapt" class="no-js" lang="en">
+<html id="adapt" class="no-js" lang="en" dir="ltr">
   <head>
     <script>
       window.ADAPT_BUILD_TYPE = '@@build.type';
+      window.ISCOOKIELMS = true;
     </script>
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
     <title></title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link href="adapt.css" type="text/css" rel="stylesheet"/>
-    <script src="libraries/js-cookie.js"></script>
-    <script src="offline_API_wrapper.js"></script>
+    <link href="adapt.css" type="text/css" rel="stylesheet">
     <script src="libraries/modernizr.js"></script>
     <script src="adapt/js/scriptLoader.js"></script>
   </head>
@@ -23,13 +22,4 @@
     </div>
   </body>
 
-  <script type="text/javascript">
-  Cookies.defaults = {
-    // uncomment this if you need the cookie to 'persist'. if left commented-out it will act as a 'session' cookie
-    // see https://github.com/js-cookie/js-cookie/tree/latest#expires
-    /* expires: 365, */
-    sameSite: 'strict'
-  }
-  </script>
-
-</html>
+  </html>
